+++ src/index.js
@@ -1,11 +1,13 @@
 import './style.css';
 // eslint-disable-next-line
 import Sortable from '../node_modules/sortablejs/modular/sortable.complete.esm';
-import Todo from './todo-class';
-import { addTodo, renderTodo } from './todo';
+import Todo from './todo-class.js';
+import { addTodo, renderTodo } from './todo.js';
+import { clearAllCompleted } from './clear-all.js';
 
 const todosListEl = document.getElementById('todos-list');
 const todoInput = document.getElementById('newtodo');
+const btnClear = document.getElementById('deleteAllCompleted');
 
 let addNewTodo = new Todo(false, todoInput.value);
 
@@ -19,6 +21,18 @@ todoInput.addEventListener('keydown', (event) => {
     todoInput.value = '';
     renderTodo();
   }
+});
+
+// eslint-disable-next-line
+btnClear.addEventListener('click', (e) => {
+  clearAllCompleted();
+});
+
+const btnRefresh = document.getElementById('refreshDom');
+// eslint-disable-next-line
+btnRefresh.addEventListener('click', (e) => {
+// eslint-disable-next-line
+  location.reload();
 });
 
 window.onload = renderTodo;
